Refactoring Footer sub-components into their own file
